requirements.txt
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
boto3==1.12.33
xlrd==1.2.0
pandas==1.0.3
requests==2.23.0
channels==2.1.1
channels-redis==2.4.0
bs4==0.0.1
gunicorn==20.0.4